scraper/requirements.txt
firecrawl-py
python-dotenv
requests
